Moving things around and renaming bat files
I want the nuget package to install the example bat files
with obvious names so that devs who are using the package
won't get confused when they see bat files with similar
names.

This is just an end-of-day savepoint. Things are in flux
and the nuget package will likely break things if it gets
published in it's current form.